Add styling to homepage details page
