+++ src/pages/HomepageDetails/HomepageDetails.css
@@ -1,0 +1,18 @@
+.bgcolor {
+  background-color: #fff;
+}
+
+.bgcolor:hover {
+  background: rgba(102, 16, 242, 0.1);
+  transition: background 2s;
+}
+
+.firstBlock {
+  padding: 10px 0px 20px 0px;
+  margin: 10px;
+}
+
+.secondBlock {
+  padding: 10px 0px 20px 0px;
+  margin: 10px;
+}
